requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
ollama>=0.3.0
chromadb>=0.4.22
pydantic>=2.5.3
streamlit>=1.30.0
reportlab>=4.0.9
sentence-transformers>=2.3.1
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.4.4
